feat: 告警处理出错给管理员发送告警信息，发送信息自己拼接，不使用模版
