+++ src/server/engine/notify.go
@@ -257,132 +257,6 @@ func notify(event *models.AlertCurEvent) {
 	}
 }
 
-// notify to maintainer to handle the error
-func notifyToMaintainer(e error, title string) {
-	event := &models.AlertCurEvent{
-		IsRecovered:  false,
-		Severity:     0,
-		RuleName:     title,
-		TriggerTime:  time.Now().Unix(),
-		TriggerValue: e.Error(),
-	}
-	maintainerUsers := memsto.UserCache.GetMaintainerUsers()
-	if len(maintainerUsers) == 0 {
-		return
-	}
-	event.NotifyUsersObj = maintainerUsers
-	event.NotifyChannelsJSON = config.C.Alerting.NotifyBuiltinChannels
-
-	LogEvent(event, "notifyToMaintainer")
-	notice := genNotice(event)
-	alertingWebhook(event)
-	handleNoticeToMaintainer(notice)
-}
-
-func handleNoticeToMaintainer(notice Notice) {
-	if len(config.C.Alerting.NotifyBuiltinChannels) == 0 {
-		return
-	}
-
-	emailset := make(map[string]struct{})
-	phoneset := make(map[string]struct{})
-	wecomset := make(map[string]struct{})
-	dingtalkset := make(map[string]struct{})
-	feishuset := make(map[string]struct{})
-
-	for _, user := range notice.Event.NotifyUsersObj {
-		if user.Email != "" {
-			emailset[user.Email] = struct{}{}
-		}
-
-		if user.Phone != "" {
-			phoneset[user.Phone] = struct{}{}
-		}
-
-		bs, err := user.Contacts.MarshalJSON()
-		if err != nil {
-			logger.Errorf("handle_notice: failed to marshal contacts: %v", err)
-			continue
-		}
-
-		ret := gjson.GetBytes(bs, "dingtalk_robot_token")
-		if ret.Exists() {
-			dingtalkset[ret.String()] = struct{}{}
-		}
-
-		ret = gjson.GetBytes(bs, "wecom_robot_token")
-		if ret.Exists() {
-			wecomset[ret.String()] = struct{}{}
-		}
-
-		ret = gjson.GetBytes(bs, "feishu_robot_token")
-		if ret.Exists() {
-			feishuset[ret.String()] = struct{}{}
-		}
-	}
-
-	phones := StringSetKeys(phoneset)
-
-	for _, ch := range config.C.Alerting.NotifyBuiltinChannels {
-		switch ch {
-		case "email":
-			if len(emailset) == 0 {
-				continue
-			}
-			subject, has := notice.Tpls["subject.tpl"]
-			if !has {
-				subject = "subject.tpl not found"
-			}
-
-			content, has := notice.Tpls["mailbody.tpl"]
-			if !has {
-				content = "mailbody.tpl not found"
-			}
-			sender.WriteEmail(subject, content, StringSetKeys(emailset))
-		case "dingtalk":
-			if len(dingtalkset) == 0 {
-				continue
-			}
-			content, has := notice.Tpls["dingtalk.tpl"]
-			if !has {
-				content = "dingtalk.tpl not found"
-			}
-			sender.SendDingtalk(sender.DingtalkMessage{
-				Title:     notice.Event.RuleName,
-				Text:      content,
-				AtMobiles: phones,
-				Tokens:    StringSetKeys(dingtalkset),
-			})
-		case "wecom":
-			if len(wecomset) == 0 {
-				continue
-			}
-			content, has := notice.Tpls["wecom.tpl"]
-			if !has {
-				content = "wecom.tpl not found"
-			}
-			sender.SendWecom(sender.WecomMessage{
-				Text:   content,
-				Tokens: StringSetKeys(wecomset),
-			})
-		case "feishu":
-			if len(feishuset) == 0 {
-				continue
-			}
-
-			content, has := notice.Tpls["feishu.tpl"]
-			if !has {
-				content = "feishu.tpl not found"
-			}
-			sender.SendFeishu(sender.FeishuMessage{
-				Text:      content,
-				AtMobiles: phones,
-				Tokens:    StringSetKeys(feishuset),
-			})
-		}
-	}
-}
-
 func alertingWebhook(event *models.AlertCurEvent) {
 	conf := config.C.Alerting.Webhook
